+++ src/components/widgets/IntakeLog/IntakeRecepientLogModal/index.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useCallback } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import actions from 'redux/companies/actions';
 import debounce from 'lodash.debounce';
-import { Form, Input, Select } from 'antd';
+import { Form, Input, Select, InputNumber } from 'antd';
 import { LoadingOutlined } from '@ant-design/icons';
 import styles from './styles.module.scss';
 
@@ -53,7 +53,6 @@ const IntakeRecepientLogModal = ({ form }) => {
       <Item
         label="Company ID"
         name="company_id"
-        className={styles.formItem}
         rules={[
           {
             required: true,
@@ -67,7 +66,6 @@ const IntakeRecepientLogModal = ({ form }) => {
       <Item
         label="Company name"
         name="company_name"
-        className={styles.formItem}
         rules={[
           {
             required: true,
@@ -82,11 +80,7 @@ const IntakeRecepientLogModal = ({ form }) => {
         />
       </Item>
 
-      <Item
-        label="Company short"
-        name="company_short"
-        className={styles.formItem}
-      >
+      <Item label="Company short" name="company_short">
         <Input
           disabled
           placeholder="Company short"
@@ -105,21 +99,41 @@ const IntakeRecepientLogModal = ({ form }) => {
           },
         ]}
       >
-        <Input type="number" placeholder="Reference pools count" />
+        <InputNumber
+          placeholder="Reference pools count"
+          className="w-100"
+          min={1}
+        />
       </Item>
 
       <Item
         label="Reference samples count"
         name="reference_samples_count"
         className={styles.formItem}
+        dependencies={['reference_pools_count']}
         rules={[
           {
             required: true,
             message: 'This field is required',
           },
+          ({ getFieldValue }) => ({
+            validator(rule, value) {
+              if (getFieldValue('reference_pools_count') <= value) {
+                return Promise.resolve();
+              }
+
+              return Promise.reject(
+                'The number of pools must be less than or equal to the number of samples',
+              );
+            },
+          }),
         ]}
       >
-        <Input type="number" placeholder="Reference samples count" />
+        <InputNumber
+          placeholder="Reference samples count"
+          className="w-100"
+          min={1}
+        />
       </Item>
 
       <Item
@@ -158,17 +172,7 @@ const IntakeRecepientLogModal = ({ form }) => {
         />
       </Item>
 
-      <Item
-        label="Tracking number"
-        name="tracking_number"
-        className={styles.formItem}
-        rules={[
-          {
-            required: true,
-            message: 'This field is required',
-          },
-        ]}
-      >
+      <Item label="Tracking number" name="tracking_number">
         <Input placeholder="Tracking Number" />
       </Item>
     </Form>