+++ src/redux/racks/actions.js
@@ -1,0 +1,9 @@
+import generateRequestActions from 'redux/factories/generateRequestActions';
+
+const actions = {
+  ...generateRequestActions(['fetch'], 'rack', 'racks'),
+  ...generateRequestActions(['get', 'update'], 'rack', 'rack'),
+  RACK_DATA_CHANGE: 'rack/RACK_DATA_CHANGE',
+};
+
+export default actions;