+++ src/redux/reducers.js
@@ -18,6 +18,7 @@ import intakeReceiptLog from './intakeReceiptLog/reducers';
 import scanSessions from './scanSessions/reducers';
 import management from './management/reducers';
 import inventory from './inventory/reducers';
+import racks from './racks/reducers';
 
 export default history =>
   combineReducers({
@@ -40,4 +41,5 @@ export default history =>
     scanSessions,
     management,
     inventory,
+    racks,
   });