+++ src/services/scans.js
@@ -15,11 +15,14 @@ export const updateTube = async ({ id, data }) => {
 };
 
 export const deleteTube = async ({ record, scanId }) => {
-  const tube = await axiosClient.delete(
-    `/scans/tubes/${record.id}/?scan_id=${scanId}`,
-  );
-
-  return tube;
+  try {
+    const tube = await axiosClient.delete(
+      `/scans/tubes/${record.id}/from/${scanId}/`,
+    );
+    return tube;
+  } catch (error) {
+    throw error;
+  }
 };
 
 export const invalidateTube = async ({ id, data }) => {